+++ LivrariaVirtual/src/main/java/br/pucminas/livrariavirtual/api/entities/Publisher.java
@@ -18,7 +18,7 @@ import br.pucminas.livrariavirtual.api.dtos.PublisherDTO;
 @Entity
 public class Publisher {
 
-	private Integer id;
+	private Long id;
 	private String name;
 	private List<Author> authors; 
 	private String about;
@@ -27,7 +27,7 @@ public class Publisher {
 	public Publisher() {}
 	
 	
-	public Publisher(Integer id, String name, List<Author> authors, String about, List<Book> books) {
+	public Publisher(Long id, String name, List<Author> authors, String about, List<Book> books) {
 		this.id = id;
 		this.name=name;
 		this.authors = authors;
@@ -37,10 +37,10 @@ public class Publisher {
 	
 	@Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
